+++ src/ThermalUnitBlock.cpp
@@ -16,8 +16,8 @@
  *         Dipartimento di Informatica \n
  *         Universita' di Pisa \n
  *
- * Copyright &copy by Antonio Frangioni, Ali Ghezelsoflu,
- *                    Rafael Durbano Lobato
+ * \copyright &copy; by Antonio Frangioni, Ali Ghezelsoflu,
+ *                     Rafael Durbano Lobato
  */
 /*--------------------------------------------------------------------------*/
 /*---------------------------- IMPLEMENTATION ------------------------------*/
@@ -47,6 +47,10 @@ using v_coeff_pair = LinearFunction::v_coeff_pair;
 // register ThermalUnitBlock to the Block factory
 
 SMSpp_insert_in_factory_cpp_1( ThermalUnitBlock );
+
+/*- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -*/
+
+int ThermalUnitBlock::f_ignore_netcdf_variables = 0;
 
 /*--------------------------------------------------------------------------*/
 /*------------------------------- FUNCTIONS --------------------------------*/
@@ -208,14 +212,17 @@ void ThermalUnitBlock::deserialize( const netCDF::NcGroup & group )
 
  ::deserialize( group , "DeltaRampUp" , v_DeltaRampUp );
  ::deserialize( group , "DeltaRampDown" , v_DeltaRampDown );
- ::deserialize( group , "PrimaryRho" , v_PrimaryRho );
- ::deserialize( group , "SecondaryRho" , v_SecondaryRho );
  ::deserialize( group , "LinearTerm" , v_LinearTerm );
  ::deserialize( group , "QuadTerm" , v_QuadTerm );
  ::deserialize( group , "ConstTerm" , v_ConstTerm );
  ::deserialize( group , "StartUpCost" , v_StartUpCost );
  ::deserialize( group , "FixedConsumption" , v_fixed_consumption );
  ::deserialize( group , "InertiaCommitment" , v_inertia_commitment );
+
+ if( ! ( f_ignore_netcdf_variables & 1 ) ) {
+  ::deserialize( group , "PrimaryRho" , v_PrimaryRho );
+  ::deserialize( group , "SecondaryRho" , v_SecondaryRho );
+ }
 
  // Decompress vectors
  decompress_vector( v_MinPower );
@@ -1077,61 +1084,39 @@ void ThermalUnitBlock::generate_abstract_constraints( Configuration * stcc ) {
  } // end( ThermalUnitBlock::generate_abstract_constraints )
 
 /*--------------------------------------------------------------------------*/
-/// verifies whether the current solution is feasible for the given constraints
-/** This function checks whether the relative violation of each RowConstraint
- * in the given group of RowConstraint is not greater than the provided
- * tolerance.
- *
- * @return This function returns true if and only if the relative violation of
- *         each RowConstraint in the given group is not greater than the given
- *         tolerance. */
-
-template<class C>
-static std::enable_if_t< std::is_base_of_v< RowConstraint , C > , bool >
-is_feasible( std::vector< C > & constraints , double tolerance ) {
- for( auto & constraint : constraints ) {
-  if( constraint.is_relaxed() )
-   continue;
-  constraint.compute();
-  if( constraint.rel_viol() > tolerance )
-   return false;
- }
- return true;
-}
-
-/*--------------------------------------------------------------------------*/
-/// verifies whether the given ColVariable are feasible
-/** This function returns true if and only if each given ColVariable is
- * feasible with respect to the given tolerance (see
- * ColVariable::is_feasible()).
- *
- * @return This function returns true if and only if each of the given
- *         ColVariable is feasible considering the given tolerance. */
-
-template<class V>
-static std::enable_if_t< std::is_base_of_v< ColVariable , V > , bool >
-is_feasible( const std::vector< V > & variables , double tolerance ) {
- for( const auto & variable : variables ) {
-  if( ! variable.is_feasible( tolerance ) )
-   return false;
- }
- return true;
-}
-
-/*--------------------------------------------------------------------------*/
 
 bool ThermalUnitBlock::is_feasible( bool useabstract , Configuration * fsbc )
 {
- // Retrieve the tolerance.
+ // Retrieve the tolerance and the type of violation.
 
- auto config = dynamic_cast< SimpleConfiguration< double > * >( fsbc );
+ double tolerance = 0;
+ bool rel_viol = true;
 
- if( ( ! config ) && f_BlockConfig )
-  config = dynamic_cast< SimpleConfiguration< double > * >
-   ( f_BlockConfig->f_is_feasible_Configuration );
+ // Try to extract, from "c", the parameters that determine feasibility.
+ // If it succeeds, it sets the values of the parameters and returns
+ // true. Otherwise, it returns false.
+ auto extract_parameters = [ & tolerance , & rel_viol ]( Configuration * c )
+  -> bool {
+  if( auto tc = dynamic_cast< SimpleConfiguration< double > * >( c ) ) {
+   tolerance = tc->f_value;
+   return true;
+  }
+  if( auto tc = dynamic_cast< SimpleConfiguration<
+      std::pair< double , int > > * >( c ) ) {
+   tolerance = tc->f_value.first;
+   rel_viol = tc->f_value.second;
+   return true;
+  }
+  return false;
+ };
 
- // If a tolerance has not been provided, use the default tolerance.
- const auto tolerance = config ? config->f_value : 1.0e-8;
+ if( ( ! extract_parameters( fsbc ) ) && f_BlockConfig )
+  // if the given Configuration is not valid, try the one from the BlockConfig
+  extract_parameters( f_BlockConfig->f_is_feasible_Configuration );
+
+ auto is_feasible = [ tolerance , rel_viol ]( auto & constraints ) {
+  return RowConstraint::is_feasible( constraints , tolerance , rel_viol );
+ };
 
  // Notice that the ZOConstraint are not checked, since the corresponding
  // check is made on the ColVariable.
@@ -1139,26 +1124,26 @@ bool ThermalUnitBlock::is_feasible( bool useabstract , Configuration * fsbc )
  return
   UnitBlock::is_feasible( useabstract )
   // Constraints
-  && ::is_feasible( Power_StartUp_ShutDown_Variables_Constraints , tolerance )
-  && ::is_feasible( Power_StartUp_Variable_Constraints , tolerance )
-  && ::is_feasible( Power_ShutDown_Variable_Constraints , tolerance )
-  && ::is_feasible( StartUp_ShutDown_Variables_Constraints , tolerance )
-  && ::is_feasible( StartUp_Constraints , tolerance )
-  && ::is_feasible( ShutDown_Constraints , tolerance )
-  && ::is_feasible( RampUp_Constraints , tolerance )
-  && ::is_feasible( RampDown_Constraints , tolerance )
-  && ::is_feasible( PrimaryRho_Constraints , tolerance )
-  && ::is_feasible( SecondaryRho_Constraints , tolerance )
-  && ::is_feasible( MinPower_Constraints , tolerance )
-  && ::is_feasible( MaxPower_Constraints , tolerance )
-  && ::is_feasible( Commitment_fixed_to_One_Constraints , tolerance )
+  && is_feasible( Power_StartUp_ShutDown_Variables_Constraints )
+  && is_feasible( Power_StartUp_Variable_Constraints )
+  && is_feasible( Power_ShutDown_Variable_Constraints )
+  && is_feasible( StartUp_ShutDown_Variables_Constraints )
+  && is_feasible( StartUp_Constraints )
+  && is_feasible( ShutDown_Constraints )
+  && is_feasible( RampUp_Constraints )
+  && is_feasible( RampDown_Constraints )
+  && is_feasible( PrimaryRho_Constraints )
+  && is_feasible( SecondaryRho_Constraints )
+  && is_feasible( MinPower_Constraints )
+  && is_feasible( MaxPower_Constraints )
+  && is_feasible( Commitment_fixed_to_One_Constraints )
   // Variables
-  && ::is_feasible( v_start_up , tolerance )
-  && ::is_feasible( v_shut_down , tolerance )
-  && ::is_feasible( v_commitment , tolerance )
-  && ::is_feasible( v_active_power , tolerance )
-  && ::is_feasible( v_primary_spinning_reserve , tolerance )
-  && ::is_feasible( v_secondary_spinning_reserve , tolerance );
+  && ColVariable::is_feasible( v_start_up , tolerance )
+  && ColVariable::is_feasible( v_shut_down , tolerance )
+  && ColVariable::is_feasible( v_commitment , tolerance )
+  && ColVariable::is_feasible( v_active_power , tolerance )
+  && ColVariable::is_feasible( v_primary_spinning_reserve , tolerance )
+  && ColVariable::is_feasible( v_secondary_spinning_reserve , tolerance );
 
  }  // end( ThermalUnitBlock::is_feasible )