+++ CHANGELOG.md
@@ -7,12 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.2] - 2023-05-17
+
+### Added
+
+- Add is_feasible() to BatteryUnitBlock, SlackUnitBlock.
+- IntermittentUnitBlock::set_BlockConfig().
+
+### Changed
+
+- Update is_feasible() in HydroUnitBlock, DCNetworkBlock,
+  IntermittentUnitBlock, ThermalUnitBlock, and BatteryUnitBlock.
+- Remove "battery_type" from BatteryUnitBlock and check if negative prices may
+  occur.
+
 ## [0.6.1] - 2022-07-01
 
 ### Added
 
 - UnitBlock can be scaled (replicated).
-- BatteryUnitBlock, IntermittentUnitBlock, and ThermalUnitBlock implement scale().
+- BatteryUnitBlock, IntermittentUnitBlock, and ThermalUnitBlock implement
+  scale().
 - BatteryUnitBlock and IntermittentUnitBlock can have their minimum and
   maximum power and storage levels scaled.
 
@@ -96,7 +111,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First test release.
 
-[Unreleased]: https://gitlab.com/smspp/ucblock/-/compare/0.6.1...develop
+[Unreleased]: https://gitlab.com/smspp/ucblock/-/compare/0.6.2...develop
+[0.6.2]: https://gitlab.com/smspp/ucblock/-/compare/0.6.1...0.6.2
 [0.6.1]: https://gitlab.com/smspp/ucblock/-/compare/0.6.0...0.6.1
 [0.6.0]: https://gitlab.com/smspp/ucblock/-/compare/0.5.0...0.6.0
 [0.5.0]: https://gitlab.com/smspp/ucblock/-/compare/0.4.1...0.5.0